+++ src/components/Button/index.js
@@ -1,5 +1,5 @@
-export { default as Button } from "./Button";
-export { StyledButton, StyledButtonLink } from "./Button.styles";
+export { default as Button } from "./Base";
+export { StyledButton, StyledButtonLink } from "./Base.styles";
 export { default as Badge } from "./Badge";
 export { default as StyledBadge } from "./Badge.styles";
 export { default as RatingBadge } from "./RatingBadge";
